pyproject.toml
[project]
name = "testproject"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "python-dotenv>=1.0.1",
    "django-allauth>=65.8.0",
    "huey>=2.5.1",
    "redis>=5.0.8",
    "pytz>=2024.1",
    "gunicorn>=23.0.0",
    "django-custom-error-views>=0.2.5",
    "psycopg[binary,pool]>=3.2.1",
    "django-storages[s3]>=1.14.6",
    "django-cotton>=0.9.32",
    "django-debug-toolbar>=5.2.0",
    "django-browser-reload>=1.18.0",
    "requests>=2.32.4",
    "django-solo>=2.4.0",
    "django-extensions>=4.1",
    "pillow>=12.1.0",
]
[tool.uv]
default-groups = "all"

[dependency-groups]
dev = [
    "django-stubs>=5.2.7",
    "pyright>=1.1.407",
    "rich>=14.2.0",
    "ruff>=0.6.3",
]

[tool.ruff]
exclude = ["onlydjango/**.*"]
